👌 Move webSocketCollection to resource dir
