+++ packages/browser-rum-core/src/boot/startRum.ts
@@ -27,7 +27,7 @@ import { startInternalContext } from '../domain/contexts/internalContext'
 import { LifeCycle, LifeCycleEventType } from '../domain/lifeCycle'
 import { startViewHistory } from '../domain/contexts/viewHistory'
 import { startRequestCollection } from '../domain/requestCollection'
-import { startWebSocketCollection } from '../domain/webSocketCollection'
+import { startWebSocketCollection } from '../domain/resource/webSocketCollection'
 import { startActionCollection } from '../domain/action/actionCollection'
 import { startErrorCollection } from '../domain/error/errorCollection'
 import { startResourceCollection } from '../domain/resource/resourceCollection'
