+++ packages/browser-rum-core/src/domain/resource/resourceCollection.ts
@@ -27,7 +27,7 @@ import { LifeCycleEventType } from '../lifeCycle'
 import type { RequestCompleteEvent } from '../requestCollection'
 import { getDocumentTraceId } from '../tracing/getDocumentTraceId'
 import { createSpanIdentifier } from '../tracing/identifier'
-import type { WebSocketCompleteEvent } from '../webSocketCollection'
+import type { WebSocketCompleteEvent } from './webSocketCollection'
 import type { GraphQlMetadata } from './graphql'
 import { extractGraphQlMetadata, findGraphQlConfiguration } from './graphql'
 import { createRequestRegistry } from './requestRegistry'
