+++ packages/browser-rum-core/src/domain/resource/webSocketCollection.spec.ts
@@ -3,11 +3,11 @@ import { initWebSocketObservable, Observable } from '@datadog/browser-core'
 import { mockClock, registerCleanupTask, type Clock } from '@datadog/browser-core/test'
 import type { ClocksState, Duration, RelativeTime } from '@datadog/js-core/time'
 import { elapsed, relativeToClocks } from '@datadog/js-core/time'
-import { mockViewHistory } from '../../test'
-import { VitalType } from '../rawRumEvent.types'
-import type { ViewHistoryEntry } from './contexts/viewHistory'
-import { LifeCycle, LifeCycleEventType } from './lifeCycle'
-import type { DurationVital } from './vital/vitalCollection'
+import { mockViewHistory } from '../../../test'
+import { VitalType } from '../../rawRumEvent.types'
+import type { ViewHistoryEntry } from '../contexts/viewHistory'
+import { LifeCycle, LifeCycleEventType } from '../lifeCycle'
+import type { DurationVital } from '../vital/vitalCollection'
 import type { WebSocketCompleteEvent } from './webSocketCollection'
 import { startWebSocketCollection, trackWebSocket, WEBSOCKET_CONNECTING_VITAL_NAME } from './webSocketCollection'
