+++ packages/browser-rum-core/src/domain/resource/webSocketCollection.ts
@@ -2,11 +2,11 @@ import type { Observable, WebSocketContext } from '@datadog/browser-core'
 import { generateUUID, initWebSocketObservable, sanitize } from '@datadog/browser-core'
 import type { ClocksState, Duration, TimeStamp } from '@datadog/js-core/time'
 import { clocksNow, elapsed } from '@datadog/js-core/time'
-import { VitalType } from '../rawRumEvent.types'
-import type { ViewHistory } from './contexts/viewHistory'
-import type { LifeCycle } from './lifeCycle'
-import { LifeCycleEventType } from './lifeCycle'
-import type { DurationVital } from './vital/vitalCollection'
+import { VitalType } from '../../rawRumEvent.types'
+import type { ViewHistory } from '../contexts/viewHistory'
+import type { LifeCycle } from '../lifeCycle'
+import { LifeCycleEventType } from '../lifeCycle'
+import type { DurationVital } from '../vital/vitalCollection'
 
 export const WEBSOCKET_CONNECTING_VITAL_NAME = 'websocket-connecting'
